pyproject.toml
[tool.poetry]
name = "fastapi-jwt"
version = "0.1.0"
description = "FastAPI jwt-token tutorial"
authors = ["ekusiadadus <ekusiadadus@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
python-jose = "^3.3.0"
mysqlclient = "^2.1.1"
SQLAlchemy = "^1.4.41"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
